refactor: CMake support, Opaque data type, compiler warnings fixed, etc.
