+++ .gitignore
@@ -1,39 +1,88 @@
-# Windows image file caches
-Thumbs.db
-ehthumbs.db
 
-# Folder config file
-Desktop.ini
+# -- IDEs --
+.vscode/
+.idea
 
-# Recycle Bin used on file shares
-$RECYCLE.BIN/
 
-# Windows Installer files
-*.cab
-*.msi
-*.msm
-*.msp
+# -- Build / CMake --
+# Compiled Object files
+*.slo
+*.lo
+*.o
+*.obj
 
-# Windows shortcuts
-*.lnk
+# Precompiled Headers
+*.gch
+*.pch
 
-# =========================
-# Operating System Files
-# =========================
+# Compiled Dynamic libraries
+*.so
+*.dylib
+*.dll
 
-# OSX
-# =========================
+# Fortran module files
+*.mod
+*.smod
 
+# Compiled Static libraries
+*.lai
+*.la
+*.a
+*.lib
+
+# Executables
+*.exe
+*.out
+*.app
+
+**/cmake-build-debug
+**/cmake-build-release
+**/CMakeCache.txt
+**/cmake_install.cmake
+**/install_manifest.txt
+**/CMakeFiles/
+**/CTestTestfile.cmake
+**/Makefile
+**/*.cbp
+**/CMakeScripts
+**/compile_commands.json
+
+include/divisible/*
+
+
+## Local
+build/**/*
+
+include/*
+lib/*
+bin/*
+test/test_runner
+
+
+
+
+# -- OS --
+# - macOS -
+# General
 .DS_Store
 .AppleDouble
 .LSOverride
 
+# Icon must end with two \r
+Icon
+
+
 # Thumbnails
 ._*
 
-# Files that might appear on external disk
+# Files that might appear in the root of a volume
+.DocumentRevisions-V100
+.fseventsd
 .Spotlight-V100
+.TemporaryItems
 .Trashes
+.VolumeIcon.icns
+.com.apple.timemachine.donotpresent
 
 # Directories potentially created on remote AFP share
 .AppleDB
