--- INSTALL
@@ -1,11 +1,0 @@
-Step 1, build dynamic shared libatomic_hash.so:
-
-cd src && make clean && make
-
-
-Step 2, install libatomic_hash.so to /usr/lib64/ and atomic_hash.h to /usr/include/
-
-make install
-
-
-Step 3, include "atomic_hash.h" in your source file(s) and dynamic link atomic_hash lib to your program. see test/makefile as example
