+++ src/hash_functions/hash_md5.h
@@ -1,0 +1,8 @@
+#ifndef HASH_MD5_H
+#define HASH_MD5_H
+
+#include <stdlib.h>
+
+void md5hash(const void *s, const size_t len, void *r);
+
+#endif /* HASH_MD5_H */
