+++ src/hash_functions/unimplemented_functions/hash_newhash.h
