--- test/atomic_hash.h
@@ -1,1 +1,0 @@
-../src/atomic_hash.h
