+++ test/utils/mtrand.c
@@ -40,10 +40,9 @@
    Vol. 8, No. 1, January 1998, pp 3--30.
 */
 
-/**************************************************
- *  Modified by Fred Huang to enable thread-safe
- *  ***********************************************
-*/
+/***************************************************
+ *  Modified by Fred Huang to enable thread-safety *
+ **************************************************/
 
 #include "mtrand.h"
 
@@ -57,9 +56,7 @@
 static unsigned long x[N];	/* the array for the state vector  */
 static unsigned long n;
 
-void
-mt_srand (unsigned long s)
-{
+void mt_srand (unsigned long s) {
   int i;
   n = 0;
   x[0] = s & MT_RAND_MAX;
@@ -67,15 +64,12 @@ mt_srand (unsigned long s)
     x[i] = (1812433253UL * (x[i-1] ^ (x[i-1] >> 30)) + i) & MT_RAND_MAX;
 }
 
-void mt_seed(void)
-{
+void mt_seed(void) {
   mt_srand (5489UL);  /* Default seed */
 }
 
 /* generates a random number on the interval [0,0xffffffff] */
-unsigned long
-mt_rand (void)
-{
+unsigned long mt_rand (void) {
   register unsigned long y, i, j;
   i = __sync_fetch_and_add (&n, 1) % N;
   j = (i == N - 1 ? 0 : i + 1);
@@ -92,16 +86,12 @@ mt_rand (void)
 }
 
 /* generates a random number on the interval [0,1]. */
-float
-mt_rand_1 (void)
-{
+float mt_rand_1 (void) {
   return ((float) mt_rand () / (float) MT_RAND_MAX);
 }
 
 /* generates a random number on the interval [0,1). */
-float
-mt_rand_lt1 (void)
-{
+float mt_rand_lt1 (void) {
   /* MT_RAND_MAX must be a float before adding one to it! */
   return ((float) mt_rand () / ((float) MT_RAND_MAX + 1.0f));
 }
